chore: address comments
OKTA-299837
