+++ src/v2/ion/RemediationConstants.js
@@ -1,4 +1,16 @@
-const FORMS = {
+/*!
+ * Copyright (c) 2019, Okta, Inc. and/or its affiliates. All rights reserved.
+ * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
+ *
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ * See the License for the specific language governing permissions and limitations under the License.
+ */
+
+ const FORMS = {
   IDENTIFY: 'identify',
   SELECT_IDENTIFY: 'select-identify',
   SELECT_ENROLL_PROFILE: 'select-enroll-profile',
