+++ src/v2/ion/ui-schema/ion-object-handler.js
@@ -1,3 +1,15 @@
+/*!
+ * Copyright (c) 2019, Okta, Inc. and/or its affiliates. All rights reserved.
+ * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
+ *
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ * See the License for the specific language governing permissions and limitations under the License.
+ */
+
 /**
  * Example of the option like
  * @param {AuthenticatorOption[]} options
@@ -8,8 +20,8 @@ const createAuthenticatorOptions = (options = []) => {
     const value = option.value && option.value.form && option.value.form.value || [];
 
     // Each authenticator option has list of ION field.
-    // Currently we are only support merely selecting one of options
-    // rather than pop up another page to collection extra data
+    // Currently we only support merely selecting one of options
+    // rather than pop up another page to collect extra data
     // (in order to fill value for `mutable: true; value: null` fields).
     // The only reason of such design is to simplify widget implementation
     // but could subject to change in later releases.
@@ -43,16 +55,24 @@ const getAuthenticatorsVerifyUiSchema = ({ options }) => {
   };
 };
 
+/**
+  * Create ui schema for ION field that has type 'object'.
+  */
 const createUiSchemaForObject = (ionFormField, remediationForm, transformedResp, createUISchema) => {
   const uiSchema = {};
 
   if (ionFormField.name === 'authenticator' && remediationForm.name === 'select-authenticator-authenticate') {
-    // similar to `factorId` but `authenticator` is a new way to model factors
-    // hence it has different structure
+    // 1. when `select-authenticator-authenticator`,
+    // create customize type so that display authenticator list as selectable list
     Object.assign(uiSchema, getAuthenticatorsVerifyUiSchema(ionFormField));
   } else if (ionFormField.name === 'authenticator' && remediationForm.name === 'select-authenticator-enroll') {
+    // 2. when `select-authenticator-enroll`,
+    // create customize type so that display authenticator list as selectable list
     Object.assign(uiSchema, getAuthenticatorsEnrollUiSchema(ionFormField));
-  } else {
+  } else if (Array.isArray(ionFormField.options)) {
+    // 3. For other cases, create ui schema for each `option` in order to render
+    // different view for each option.
+    //
     // e.g. { "name": "credentials", "type": "object", options: [ {value: {form: value:[]} ]
     uiSchema.optionsUiSchemas = ionFormField.options.map(opt => {
       return createUISchema(transformedResp, {
